[next-config] Organized more files
