+++ client/src/@types/Course/Course.ts
