+++ client/src/@types/Course/MockCourse.ts
