+++ client/src/@types/Login/LoginPageReducerAction.ts
@@ -1,0 +1,10 @@
+import type { LoginPageReducerActionType } from "./LoginPageReducerActionType";
+import type { LoginPageState } from "./LoginPageState";
+
+/**
+ * Action that will be utilized in the reducer
+ */
+export type LoginPageReducerAction = {
+    type: LoginPageReducerActionType;
+    payload: LoginPageState;
+};
