+++ client/src/@types/Login/LoginPageReducerActionType.ts
@@ -1,0 +1,8 @@
+/**
+ * Types of actions the user can make
+ */
+export type LoginPageReducerActionType =
+    | "setLoginOverlay"
+    | "setPasswordOverlay"
+    | "setShowPassword"
+    | "setSignUpOverlay";
