+++ client/src/@types/Login/LoginPageState.ts
@@ -1,0 +1,9 @@
+/**
+ * Interface for managing the state of overlays
+ */
+export type LoginPageState = {
+    showPasswordOverlay: boolean;
+    showLoginOverlay: boolean;
+    showSignUpOverlay: boolean;
+    showPassword: boolean;
+};
