+++ client/src/@types/index.ts
@@ -1,7 +1,19 @@
 import type { LoginResponse } from "./api/LoginResponse";
-import type { Course } from "./Course";
-import type { MockCourse } from "./MockCourse";
-import type { Plan } from "./Plan";
-import type { Semester } from "./Semester";
+import type { Course } from "./Course/Course";
+import type { MockCourse } from "./Course/MockCourse";
+import type { LoginPageReducerAction } from "./Login/LoginPageReducerAction";
+import type { LoginPageReducerActionType } from "./Login/LoginPageReducerActionType";
+import type { LoginPageState } from "./Login/LoginPageState";
+import type { Plan } from "./Plan/Plan";
+import type { Semester } from "./Semester/Semester";
 
-export type { Course, LoginResponse, MockCourse, Plan, Semester };
+export type {
+    Course,
+    LoginPageReducerAction,
+    LoginPageReducerActionType,
+    LoginPageState,
+    LoginResponse,
+    MockCourse,
+    Plan,
+    Semester,
+};
