+++ client/src/reducer/index.ts
@@ -1,0 +1,3 @@
+import { LoginPageReducer } from "./LoginPageReducer/LoginPageReducer";
+
+export { LoginPageReducer };
